+++ lib/src/flutter_local_notifications.dart
@@ -6,7 +6,7 @@ typedef Future<dynamic> SelectNotificationCallback(String payload);
 /// Signature of callback passed to [initialize].
 /// Callback is triggered when a notification is shown. Note that on iOS this only works on iOS 10+.
 /// The callback must be a top-level or static method
-typedef Future<dynamic> NotificationCallback(
+typedef Future<dynamic> ShowNotificationCallback(
     int id, String title, String body, String payload);
 
 /// The available intervals for periodically showing notifications
@@ -78,7 +78,7 @@ class FlutterLocalNotificationsPlugin {
   /// Initializes the plugin. Call this method on application before using the plugin further
   Future<bool> initialize(InitializationSettings initializationSettings,
       {SelectNotificationCallback onSelectNotification,
-      NotificationCallback onNotification}) async {
+      ShowNotificationCallback onShowNotification}) async {
     onSelectNotification = onSelectNotification;
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificInitializationSettings(initializationSettings);
@@ -91,11 +91,11 @@ class FlutterLocalNotificationsPlugin {
     var headlessInitializationSettings = <String, dynamic>{
       'callbackDispatcher': callback.toRawHandle()
     };
-    if (onNotification != null) {
+    if (onShowNotification != null) {
       headlessInitializationSettings['onNotificationCallbackDispatcher'] =
-          PluginUtilities.getCallbackHandle(onNotification).toRawHandle();
+          PluginUtilities.getCallbackHandle(onShowNotification).toRawHandle();
     }
-    print(result.toString());
+
     await _channel.invokeMethod(
         'initializeHeadlessService', headlessInitializationSettings);
     return result;
@@ -111,6 +111,7 @@ class FlutterLocalNotificationsPlugin {
   Future show(int id, String title, String body,
       NotificationDetails notificationDetails,
       {String payload}) async {
+    _validateId(id);
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificNotificationDetails(notificationDetails);
     await _channel.invokeMethod('show', <String, dynamic>{
@@ -124,6 +125,7 @@ class FlutterLocalNotificationsPlugin {
 
   /// Cancel/remove the notification with the specified id. This applies to notifications that have been scheduled and those that have already been presented.
   Future cancel(int id) async {
+    _validateId(id);
     await _channel.invokeMethod('cancel', id);
   }
 
@@ -136,6 +138,7 @@ class FlutterLocalNotificationsPlugin {
   Future schedule(int id, String title, String body, DateTime scheduledDate,
       NotificationDetails notificationDetails,
       {String payload}) async {
+    _validateId(id);
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificNotificationDetails(notificationDetails);
     await _channel.invokeMethod('schedule', <String, dynamic>{
@@ -153,6 +156,7 @@ class FlutterLocalNotificationsPlugin {
   Future periodicallyShow(int id, String title, String body,
       RepeatInterval repeatInterval, NotificationDetails notificationDetails,
       {String payload}) async {
+    _validateId(id);
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificNotificationDetails(notificationDetails);
     await _channel.invokeMethod('periodicallyShow', <String, dynamic>{
@@ -170,6 +174,7 @@ class FlutterLocalNotificationsPlugin {
   Future showDailyAtTime(int id, String title, String body,
       Time notificationTime, NotificationDetails notificationDetails,
       {String payload}) async {
+    _validateId(id);
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificNotificationDetails(notificationDetails);
     await _channel.invokeMethod('showDailyAtTime', <String, dynamic>{
@@ -188,6 +193,7 @@ class FlutterLocalNotificationsPlugin {
   Future showWeeklyAtDayAndTime(int id, String title, String body, Day day,
       Time notificationTime, NotificationDetails notificationDetails,
       {String payload}) async {
+    _validateId(id);
     var serializedPlatformSpecifics =
         _retrievePlatformSpecificNotificationDetails(notificationDetails);
     await _channel.invokeMethod('showWeeklyAtDayAndTime', <String, dynamic>{
@@ -227,5 +233,12 @@ class FlutterLocalNotificationsPlugin {
 
   Future _handleMethod(MethodCall call) {
     return onSelectNotification(call.arguments);
+  }
+
+  void _validateId(int id) {
+    if (id > 0x7FFFFFFF || id < -0x80000000) {
+      throw ArgumentError(
+          'id must fit within the size of a 32-bit integer i.e. in the range [-2^31, 2^31 - 1]');
+    }
   }
 }